pyproject.toml
[project]
name = "dra"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "openai-agents>=0.2.9",
    "pydantic>=2.11.7",
    "pyfiglet>=1.0.4",
    "rich>=14.1.0",
    "tavily-python>=0.7.11",
]
